+++ src/capacitor-run.ts
@@ -12,7 +12,7 @@ import { npx, PackageManager, preflightNPMCheck } from './node-commands';
 import { Project } from './project';
 import { gradleToJson } from './gradle-to-json';
 import { ExtensionSetting, getExtSetting, getSetting, WorkspaceSetting } from './workspace-state';
-import { workspace } from 'vscode';
+import { window, workspace } from 'vscode';
 import { join } from 'path';
 
 /**
@@ -21,7 +21,7 @@ import { join } from 'path';
  * @param  {Project} project
  * @returns string
  */
-export function capacitorRun(project: Project, platform: CapacitorPlatform): string {
+export async function capacitorRun(project: Project, platform: CapacitorPlatform): Promise<string> {
   let preop = '';
   let rebuilt = false;
   let noSync = false;
@@ -47,7 +47,7 @@ export function capacitorRun(project: Project, platform: CapacitorPlatform): str
     case MonoRepoType.yarn:
     case MonoRepoType.lerna:
     case MonoRepoType.npm:
-      return preop + capRun(platform, project.repoType, rebuilt, noSync, project);
+      return preop + (await capRun(platform, project.repoType, rebuilt, noSync, project));
     case MonoRepoType.nx:
       return preop + nxRun(platform, project.repoType, rebuilt, noSync, project);
     default:
@@ -67,13 +67,13 @@ export function useIonicCLI(): boolean {
   return exists('@ionic/cli');
 }
 
-function capRun(
+async function capRun(
   platform: CapacitorPlatform,
   repoType: MonoRepoType,
   noBuild: boolean,
   noSync: boolean,
   project: Project,
-): string {
+): Promise<string> {
   let liveReload = getSetting(WorkspaceSetting.liveReload);
   const externalIP = !getExtSetting(ExtensionSetting.internalAddress);
   const httpsForWeb = getSetting(WorkspaceSetting.httpsForWeb);
@@ -121,8 +121,8 @@ function capRun(
 
   capRunFlags += getConfigurationArgs();
 
-  const flavors = getFlavors(platform, project);
-  if (!flavors) return;
+  const flavors = await getFlavors(platform, project);
+  if (flavors == undefined) return;
   capRunFlags += flavors;
 
   capRunFlags += InternalCommand.publicHost;
@@ -152,15 +152,15 @@ function capRun(
   } ${capRunFlags}`;
 }
 
-function nxRun(
+async function nxRun(
   platform: CapacitorPlatform,
   repoType: MonoRepoType,
   noBuild: boolean,
   noSync: boolean,
   project: Project,
-): string {
+): Promise<string> {
   if (project.monoRepo?.isNXStandalone) {
-    return capRun(platform, repoType, noBuild, noSync, project);
+    return await capRun(platform, repoType, noBuild, noSync, project);
   }
   // Note: This may change, see: https://github.com/nxtend-team/nxtend/issues/490
   return `${npx(project.packageManager)} nx run ${project.monoRepo.name}:cap --cmd "run ${platform} --target=${
@@ -168,19 +168,28 @@ function nxRun(
   }"`;
 }
 
-function getFlavors(platform: CapacitorPlatform, prj: Project): string | undefined {
+async function getFlavors(platform: CapacitorPlatform, prj: Project): Promise<string | undefined> {
   if (platform == CapacitorPlatform.ios) {
     return '';
   }
 
-  const buildGradle = join(prj.projectFolder(), 'android', 'app', 'build.gradle');
-  const data = gradleToJson(buildGradle);
-  if (data?.android?.productFlavors) {
-    const list = Object.keys(data.android.productFlavors);
-    if (list?.length == 0) {
-      return '';
+  if (ionicState.flavors == undefined) {
+    ionicState.flavors = [];
+    const buildGradle = join(prj.projectFolder(), 'android', 'app', 'build.gradle');
+    const data = gradleToJson(buildGradle);
+    if (data?.android?.productFlavors) {
+      const list = Object.keys(data.android.productFlavors);
+      if (list?.length == 0) {
+        return '';
+      }
+      ionicState.flavors = list;
     }
-    return list[0];
   }
-  return '';
+  if (ionicState.flavors.length == 0) {
+    return '';
+  }
+
+  const selection = await window.showQuickPick(ionicState.flavors, { placeHolder: 'Select the Android Flavor to run' });
+  if (!selection) return undefined;
+  return ` --flavor=${selection}`;
 }